requirements.txt
Flask==3.0.2
requests==2.31.0
gunicorn==22.0.0
python-dotenv==1.0.1
defusedxml==0.7.1